Initial commit of pre2000 v2.0 files
These are copies of the v2.0 files with the id updated to include "_pre2000" and the start and end date adjusted.

Should the version parameter include "_pre2000", or is it okay to leave as 2.0?

I set the start and end to be the water years (1984-10-01 to 1999-10-01) but we may want to limit the start to be 1985 and/or include the end of 1999 so that a user wouldn't need to merge the collections if they wanted just 1999.