Moving catalog and description files out of v2_0 folder
I'm not sure what the exact structure of this repo should be, but we might want to move everything into an "OpenET" folder to match the public data catalog structure.

None of the other datasets in the public data catalog have subfolders, but they seem useful for keeping us organized so it might be worth keeping them for now.